De-duplicator should send previous data point on change
This fixes an issue where we'd de-duplicate data and not send it and then when the
data did change the graph would end up with the incorrect slope.  Now we store the
previous data point and send it before we send the changed data point, so we only
de-dupe the points in the middle of a stream.

Change-Id: I0788b71e351cad888702cd4421e1c9b36503d92c
Reviewed-on: https://review.stumble.net/305
Tested-by: Dave Barr <barr@stumbleupon.com>
Reviewed-by: Dave Barr <barr@stumbleupon.com>