Threaded + stdin tcollector
This major change to the way tcollector works should make the entire thing be
safe for the collectors.  We should always be reading input even if we're not
able to send it off to the TSD.

This also gives us exponential backoff of the connection establishment to the
TSD so we shouldn't slam it so hard when it dies.

Change-Id: Ifba1da1ee6482f7946b7edba38d05b2ad262263f