change tcollector timeout to 15 seconds
Have tcollector wait longer for tsd to respond.  When tsd is loaded
it can take longer than 5 seconds for it to respond to version/etc.
This is causing tcollector to close the connection, thinking it's dead.
This is happening a lot, causing tcollector to back off and get further
behind.

Change-Id: I18595fbcfc18bc9c65b58a7c341a5747b8add2b4
Reviewed-on: https://review.stumble.net/518
Tested-by: Mark Smith <msmith@stumbleupon.com>
Reviewed-by: Mark Smith <msmith@stumbleupon.com>