Add tcollector metrics and verify less
This reduces the connection verification code so that it only verifies that the
TSD is up and running every 60 seconds.  Worst case, this means we could lose up
to a minute worth of data if it goes down.  That's acceptable.

This also adds tcollector specific metrics so that we can track how much data
is being sent in by the collectors, how much we're dropping, etc.

Change-Id: Ie490ce71071ee58c9601bebdd84d2311f923bf02
Reviewed-on: https://review.stumble.net/524
Reviewed-by: Benoit Sigoure <tsuna@stumbleupon.com>
Tested-by: Dave Barr <barr@stumbleupon.com>
Reviewed-by: Dave Barr <barr@stumbleupon.com>