tcollector startstop: try to reduce restarts
Try to reduce the number of restarts of tcollector, partly because
of the brokenness of how puppet insists on refreshing, even when
things like the mtime of directories changes because of actions
that took place during the previous puppet run, which causes
tcollector to be restarted twice after an update.  Instead, let's
just restart tcollector when tcollector.py itself has changed.
We could also include something like a max-uptime we'll allow
tcollector to be, but for now this is better than these constant
unnecessary restarts.

Change-Id: If3517f92cc4312c52b13e3f8d97c19b52edd213a
Reviewed-on: https://review.stumble.net/537
Tested-by: Dave Barr <barr@stumbleupon.com>
Reviewed-by: Dave Barr <barr@stumbleupon.com>