Fix a deadlock during graceful shutdown.
The SenderThread is waiting (blocking) on an Event owned by the
ReaderThread.  When the ReaderThread has new data, it set()s the
Event to wake the SenderThread.  If there's no data to send and
the ReaderThread exits (because tcollector is asked to shutdown,
for instance), the SenderThread will stick around forever, blocked
on an Event that will never be set().

There are 2 ways of solving this problem:
  1. Have the SenderThread wait() on the Event with a timeout.
     Whenever wait() times out, check if the thread is supposed
     to exit.
  2. Have ReaderThread set() the Event before exiting to wake
     the SenderThread.  The SenderThread will do one useless
     iteration and then realize that it needs to exit.

I chose the 2nd option, because it's simpler and improves the
"liveness" of the application when shutting down (because the
SenderThread gets notified immediately when the ReaderThread
exits, instead of waiting for a timeout to happen).

Change-Id: I9a7f4d3ebee25610119ce3867b4c01553f1dc0ba
Reviewed-on: https://review.stumble.net/676
Tested-by: Benoit Sigoure <tsuna@stumbleupon.com>
Reviewed-by: Dave Barr <barr@stumbleupon.com>