tcollector.py Rewrite reader queue to use Queue()
This ends up completely redoing the SenderThread and ReaderThread
scheduling.  Instead of doing locks and events back and forth to
block/wake up the thread, the ReaderThread just reads stuff from
collectors and puts data into its reader Queue, which the SenderThread
waits on and consumes.  This gets rid of the old tempq array queue.

Instead of doing the whole "queue too big", just let Queue's maxsize
do the work for us.  I create a subclass of Queue for this, which
implements a non-blocking put() that just logs when we are throwing
away data.

Rename the 'outq' mini send_data array queue to 'sendq'.

Change-Id: I0480a9cf6ae32cc3d7fd037c39266cf2da68ee37
Reviewed-on: https://review.stumble.net/682
Reviewed-by: Benoit Sigoure <tsuna@stumbleupon.com>
Tested-by: Dave Barr <barr@stumbleupon.com>