Fix restart in case pidfile is stale
If the pidfile is stale, tcollector won't get restarted
if the pid got reused.

Don't do simplistic checks of the pid running but just
always use the pgrep check.

Removes blindly using the pidfile in stop() too.

Should just get rid of using the pidfile, but currently
it's required as part of tcollector.py.  The remaining
use case is the forcerestart logic, but we could get
the same functionality from looking at start time from
/proc/PID

Change-Id: I24b3fa0eb10332cac51e4d28a06ae030147b2c28
Reviewed-on: https://review.stumble.net/2236
Reviewed-by: Benoit Sigoure <tsuna@stumbleupon.com>
Tested-by: Benoit Sigoure <tsuna@stumbleupon.com>