Evict old keys from the de-dup cache.
For every combination of (metric, tags), collectors remember what was
the last value they saw so that they can remove duplicate values (or,
in the future, perform RLE encoding).  If there are loads of different
combinations of (metric, tags) changing over time, this can lead to
excessive memory consumption because old values are never evicted from
this "de-dup cache".

This change adds a flag, --evict-interval (set to 6000 seconds = 1h40m
by default), to put an upper bound on how long a collector will remember
the last value seen for a specific combination of (metric, tags).

This change is based on a contribution of Kai Ren <kair at cs.cmu.edu>.

Change-Id: I5b5b0b64e05b2f1f81b213e14117e029bfd25ba7