Add a collector for MySQL.
The collector includes about 300 metrics about MySQL (when InnoDB
is used).  Most metrics are collected through `SHOW GLOBAL STATUS'.

The collector has a configuration file, `mysqlconf.py', in which
the user / password to use to connect to MySQL must be specified.

The collector has limited support for MySQL 5.0, because in that
version of MySQL running the command `SHOW GLOBAL STATUS' has a
big impact on the performance of the database.  Hopefully almost
everyone uses at least MySQL 5.1 these days.

Change-Id: If6042427f1701da3f5954c166a087c7996a86a1f