Add Redis collector
This collector gathers data from local Redis servers.  This requires the Redis module for Python.  We use netstat to look for 'redis-server' processes running on the local machine, since many people run multiple Redis servers per box.

It is also suggested you put a hint in your Redis configuration file to tell this collector a logical 'cluster' name.  This helps if you have several Redis instances on different hosts and you want to be able to aggregate the data.