Add Riak collector
This collector is for the Riak distributed database.  It uses the stats JSON object to parse out data and create some timeseries.

This expects /usr/lib/riak to exist (it does by default) and it uses the default ports.  This expects you to only be running one Riak instance on a machine.  This also only collects stats from the local machine -- you will need to run collectors on every machine you use as a Riak node.