Ignore the per-table/region/cf metrics for now.
Storing them as-is in TSDB causes the creation of hundreds of metrics
and tags, which isn't useful as-is.  The collector would need to do a
bit of aggregation and clean up to properly present this data to TSDB.
Until the collector is improved, just drop these metrics to avoid the
ton of mostly unusable noise they create.