More robust HAProxy collector with simpler config
The HAProxy collector is made more robust in the face of possible
HAProxy output changes by using a proper CSV parsing library, rather
than relying on indices in a list.

It is easier to configure for additional values that one may wish to
report by simply providing the names of the metrics to report as a list,
defined at the top of the file.