Add option for specifying HTTP endpoint URL path (#359)
* Make logged URL consistent with actual used URL

The "details" query portion would always be present in logged URL
regardless of whether it would actually be used.

Tidy up the URL generation a little bit to make it easier to keep
consistent.

* Factor URL-building out into its own little method

* Add option for specifying the HTTP endpoint URL path

The option can be useful for cases where you need multiple tcollector
instances to report to different URLs to be able to differentiate between
them without having to add redundant data (via e.g. the "tag" option for
tcollector). In our use case we're proxying the requests to different TSD
instances.