eos/collectors: sum up multiple instances of each agent
When an agent has multiple processes, aggregate their stats together
upfront to make things easier to graph.  We could also report a break
down per instance but it's hard to tag these properly.  Typically there
would be one per VRF or per Linecard, but we don't really have a generic
way of extracting an appropriate tag, and using the PID as a substitute
isn't really useful because it's impossible in hindsight to graph things
properly as we don't know which pid was used for each instance at any
given point in the past.