Fix order of checking if value is bool
bool is subclass of int and will match against a check in
`utils.is_numeric()`. This commit checks if `v` is bool before anything
else, so we catch it early.

```
>>> int.__subclasses__()
[<type 'bool'>]
>>> isinstance(True, (int))
True
>>> isinstance(True, (bool))
True
```

Previous behaviour was that jolokia tried to send `True` for booleans as
a value to opentsdb, which isn't allowed, and caused
`java.lang.NumberFormatException: For input string: "True"` in
application logs.