fix elasticsearch sending booleans to opentsdb
Added a generic bool-to-int convert in printmetric-function, so that any
booleans sent for printing will be converted to an integer.
Fixes an an issue where the metric `is_master` tried to send boolean
values.

If this is too generic, one could fix the specific `is_master` value in
`_collect_server`-function, line
[185](https://github.com/OpenTSDB/tcollector/blob/master/collectors/0/elasticsearch.py#L185)
or
[187](https://github.com/OpenTSDB/tcollector/blob/master/collectors/0/elasticsearch.py#L187)
by explicitly changing only that to `int()`, but figured the generic
approach is future-proof for any other metrics that might slip through
as booleans.