Fix order of checking if value is bool (#428)
* Fix order of checking if value is bool

bool is subclass of int and will match against a check in
`utils.is_numeric()`. This commit checks if `v` is bool before anything
else, so we catch it early.

```
>>> int.__subclasses__()
[<type 'bool'>]
>>> isinstance(True, (int))
True
>>> isinstance(True, (bool))
True
```

Previous behaviour was that jolokia tried to send `True` for booleans as
a value to opentsdb, which isn't allowed, and caused
`java.lang.NumberFormatException: For input string: "True"` in
application logs.

* added characters to replace/remove

jolokia has some metrics with single-tics in them as such:

`CodeHeap 'non-profiled nmethods'":
{"init":2555904,"committed":18087936,"max":121737216,"used":14709504}`

This would translate into invalid metric-names. Opentsdb accepts this
regexp in metric/tags: `[-_./a-zA-Z0-9]+`

Also changed the single-line .replace().replace() string operation into
separate lines for readability.